requirements.txt
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.4
cssselect==1.2.0
fastjsonschema==2.19.1
filelock==3.13.1
future==1.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
listcrunch==1.0.1
lxml==5.1.0
packaging==23.2
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-documentcloud==4.2.0
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
ratelimit==2.2.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
tldextract==5.1.1
Twisted==23.10.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.2
